PFP-class

An S4 object for storing pathway fingerprint scores information.

Slots

pathways_score,

a list contains PFP_score, stats_test, random_score, genes_score. PFP_score is a numeric score indicating the performance of a gene_list in some pathways. stats_test is a statistic test for the PFP_score. random_score is used for the statistic test. genes_score is the detail scores of every gene in the gene_list.

refnet_info,

a data.frame, which contains the specific information of pathway networks. Just be the same as net_info in PFP-class, including the index, id, name, group and species.

method

pathways_score, signature(object = "PFP"): extract the pathways score

refnet_info, signature(object = "PFP"): extract the pathway networks information

PFP_score, signature(object = "PFP"): extract the PFP score

stats_test, signature(object = "PFP"): extract p_value & p_adj_value

random_score, signature(object = "PFP"): extract the random score

genes_score, signature(object = "PFP", index=NULL, index_type = c("pathway_id","pathway_name","slice")): extract the genes score

refnet_names, signature(object = "PFP"): extract the refnet names

sub_PFP, signature(object = "PFP", group_name = NULL, index = NULL, index_type = c("slice","pathway_id","pathway_name")) subset of PFP object

show_PFP, signature(object = "PFP"): display methods for S4 classes PFP

plot_PFP, signature(object, type = "character", p_size = "numeric", l_size = 'numeric'): plot the Pathway Fingerprint.

rank_PFP signature(object = "PFP", total_rank = FALSE, decreasing=TRUE) sort the PFP score.

See also